+++ example/lib/pages/dialog.dart
@@ -1,3 +1,4 @@
+import 'package:awesome_flutter_extensions/awesome_flutter_extensions.dart';
 import 'package:example/common/base_scaffold.dart';
 import 'package:flutter/material.dart';
 import 'package:shadcn_ui/shadcn_ui.dart';
@@ -51,7 +52,7 @@ class DialogPage extends StatelessWidget {
                             ],
                           ),
                         )
-                        .toList(),
+                        .separatedBy(const SizedBox(height: 16)),
                   ),
                 ),
               ),